shapeTracker/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
contourpy==1.1.0
cycler==0.11.0
docopt==0.6.2
Flask==2.3.2
fonttools==4.41.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mysql-connector-python==8.0.33
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.3
Werkzeug==2.3.6
yarg==0.1.9
